🚚 extract ReactionIcon
