+++ app/web/wayforge-client/src/components/energy/EnergyCard_B.tsx
@@ -1,24 +1,15 @@
 import type { FC } from "react"
 
 import { css } from "@emotion/react"
-import styled from "@emotion/styled"
-import corners, { chamfer } from "corners"
 import { useRecoilValue } from "recoil"
 
 import { ListItems } from "~/app/web/wayforge-client/recoil-list"
 import type { RecoilListItemProps } from "~/app/web/wayforge-client/recoil-list"
 import { Luum } from "~/packages/Luum/src"
 
-import {
-  Span_EnergyAmount,
-  Span_VoidIcon,
-  SVG_EnergyIcon,
-} from "./EnergyIcon_SVG"
-import {
-  findEnergyWithRelationsState,
-  findEnergyState,
-} from "../../services/energy"
-import { findReactionEnergyState } from "../../services/energy_reaction"
+import { SVG_EnergyIcon } from "./EnergyIcon"
+import { ReactionIcon_INTERNAL } from "./ReactionIcon"
+import { findEnergyWithRelationsState } from "../../services/energy"
 import type { Reaction, ReactionRelations } from "../../services/reaction"
 import { findReactionWithRelationsState } from "../../services/reaction"
 import { cssCard } from "../Card"
@@ -33,31 +24,10 @@ export function writePathPoint(
   return command ? `${command} ${x},${y}` : `  ${x},${y}`
 }
 
-const energyListCss = css`
-  flex-grow: 1;
-  width: 50%;
-  height: 100%;
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  padding: 2px;
-  gap: 1px;
-`
-
 export const Div_EnergyCardFeature: FC<
   RecoilListItemProps<Reaction & ReactionRelations>
 > = ({ label, findState }) => {
   const reaction = useRecoilValue(findState(label.id))
-  const energy = useRecoilValue(findReactionEnergyState(reaction.id))
-  const colorB = Luum.fromJSON(energy.colorB)
-  const energyPresentHex = colorB.tint(10).hex
-  const energyAbsentHex = colorB.shade(10).hex
-  const doesProduceEnergy = reaction.products.some(
-    (product) => product.id === energy.id
-  )
-  const doesConsumeEnergy = reaction.reagents.some(
-    (reagent) => reagent.id === energy.id
-  )
 
   return (
     <div
@@ -94,86 +64,7 @@ export const Div_EnergyCardFeature: FC<
           {reaction.timeUnit}
         </small>
       </h2>
-      <div>
-        <ListItems
-          findState={findEnergyState}
-          labels={reaction.reagents}
-          Components={{
-            Wrapper: styled(
-              corners(null, null, chamfer, null).options({
-                cornerSize: 5,
-                noClipping: true,
-                below: {
-                  color: doesConsumeEnergy ? energyPresentHex : energyAbsentHex,
-                  stroke: {
-                    color: energyAbsentHex,
-                    width: 1,
-                  },
-                },
-              }).span
-            )(energyListCss),
-            ListItem: Span_EnergyAmount,
-            NoItems: () => (
-              <Span_VoidIcon
-                size={15}
-                colorA={colorB.tint(20)}
-                colorB={colorB}
-              />
-            ),
-          }}
-        />
-        <svg
-          width="18"
-          height="18"
-          viewBox="0 0 12 12"
-          style={{
-            marginRight: -9,
-            marginLeft: -9,
-            zIndex: 1,
-          }}
-        >
-          <path
-            d={[
-              writePathPoint(2, 4, `M`),
-              writePathPoint(5, 4, `L`),
-              writePathPoint(5, 0, `L`),
-              writePathPoint(10, 6, `L`),
-              writePathPoint(5, 12, `L`),
-              writePathPoint(5, 8, `L`),
-              writePathPoint(2, 8, `L`),
-              `Z`,
-            ].join(` `)}
-            fill={colorB.shade(5).hex}
-            stroke={energyPresentHex}
-          />
-        </svg>
-        <ListItems
-          findState={findEnergyState}
-          labels={reaction.products}
-          Components={{
-            Wrapper: styled(
-              corners(null, null, null, null).options({
-                noClipping: true,
-                below: {
-                  color: doesProduceEnergy ? energyPresentHex : energyAbsentHex,
-                  stroke: {
-                    color: energyAbsentHex,
-                    width: 1,
-                  },
-                },
-              }).span
-            )(energyListCss),
-            ListItem: Span_EnergyAmount,
-            NoItems: () => (
-              <Span_VoidIcon
-                size={15}
-                colorA={colorB.tint(20)}
-                colorB={colorB}
-              />
-            ),
-          }}
-        />
-      </div>
+      <ReactionIcon_INTERNAL reaction={reaction} size={20} />
     </div>
   )
 }
