+++ app/web/wayforge-client/src/components/energy/EnergyListItem.tsx
@@ -5,7 +5,7 @@ import { useNavigate } from "react-router-dom"
 import type { RecoilListItemProps } from "~/app/web/wayforge-client/recoil-list"
 import { RecoverableErrorBoundary } from "~/packages/hamr/react-ui/error-boundary"
 
-import { SVG_EnergyIcon } from "./EnergyIcon_SVG"
+import { SVG_EnergyIcon } from "./EnergyIcon"
 import type { Energy } from "../../services/energy"
 
 export const EnergyListItem: FC<RecoilListItemProps<Energy>> = ({ label }) => {
