+++ app/web/wayforge-client/src/components/energy/ReactionHome.tsx
@@ -1,6 +1,5 @@
 import type { FC } from "react"
 
-import { css } from "@emotion/react"
 import styled from "@emotion/styled"
 import { useRecoilValue } from "recoil"
 
@@ -20,32 +19,28 @@ export const ReactionHome: FC = () => {
   const ids = useRecoilValue(reactionIndex)
 
   return (
-    <div
-      css={css`
-        border: 2px solid #333;
-      `}
-    >
-      <ListItems
-        labels={[...ids].map((id) => ({ id }))}
-        findState={findReactionState}
-        useCreate={useAddReaction}
-        useRemove={useRemoveReaction}
-        Components={{
-          Wrapper: styled.ul`
-            padding: 20px;
-            display: flex;
-            flex-wrap: wrap;
-            list-style: none;
-            list-style-type: none;
-          `,
-          ListItemWrapper: ({ children }) => <li>{children}</li>,
-          ListItem: ReactionListItem,
-          ItemCreator: ({ useCreate }) => {
-            const create = useCreate()
-            return <button onClick={create}>Add</button>
-          },
-        }}
-      />
-    </div>
+    <ListItems
+      labels={[...ids].map((id) => ({ id }))}
+      findState={findReactionState}
+      useCreate={useAddReaction}
+      useRemove={useRemoveReaction}
+      Components={{
+        Wrapper: styled.ul`
+          padding: 20px;
+          display: inline-flex;
+          flex-wrap: wrap;
+          list-style: none;
+          list-style-type: none;
+          border: 2px solid #333;
+          gap: 5px;
+        `,
+        ListItemWrapper: ({ children }) => <li>{children}</li>,
+        ListItem: ReactionListItem,
+        ItemCreator: ({ useCreate }) => {
+          const create = useCreate()
+          return <button onClick={create}>Add</button>
+        },
+      }}
+    />
   )
 }
